Move XXXPool from QMCApp to corresponding folders.
