+++ src/QMCDrivers/tests/test_ContextForSteps.cpp
@@ -13,9 +13,9 @@
 
 #include "Configuration.h"
 #include "QMCDrivers/ContextForSteps.h"
-#include "QMCApp/tests/MinimalWaveFunctionPool.h"
-#include "QMCApp/tests/MinimalParticlePool.h"
-#include "QMCApp/tests/MinimalHamiltonianPool.h"
+#include "QMCWaveFunctions/tests/MinimalWaveFunctionPool.h"
+#include "Particle/tests/MinimalParticlePool.h"
+#include "QMCHamiltonians/tests/MinimalHamiltonianPool.h"
 
 #include "OhmmsPETE/TinyVector.h"
 #include "OhmmsPETE/TinyVectorOps.h"
