+++ src/QMCDrivers/tests/test_Crowd.cpp
@@ -17,9 +17,9 @@
 #include "type_traits/template_types.hpp"
 #include "Estimators/tests/FakeEstimator.h"
 
-#include "QMCApp/tests/MinimalWaveFunctionPool.h"
-#include "QMCApp/tests/MinimalParticlePool.h"
-#include "QMCApp/tests/MinimalHamiltonianPool.h"
+#include "QMCWaveFunctions/tests/MinimalWaveFunctionPool.h"
+#include "Particle/tests/MinimalParticlePool.h"
+#include "QMCHamiltonians/tests/MinimalHamiltonianPool.h"
 
 #include "QMCDrivers/tests/SetupPools.h"
