+++ src/QMCWaveFunctions/tests/CMakeLists.txt
@@ -56,12 +56,12 @@ ENDIF()
 
 ADD_EXECUTABLE(${UTEST_EXE} test_wf.cpp test_TrialWaveFunction.cpp test_wavefunction_factory.cpp
                test_bspline_jastrow.cpp test_counting_jastrow.cpp test_polynomial_eeI_jastrow.cpp
-               test_einset.cpp test_hybridrep.cpp test_pw.cpp
+               test_einset.cpp test_hybridrep.cpp test_pw.cpp test_wavefunction_pool.cpp
                test_dirac_det.cpp test_multi_dirac_determinant.cpp test_dirac_matrix.cpp
                test_rpa_jastrow.cpp test_example_he.cpp test_user_jastrow.cpp
                test_lattice_gaussian.cpp test_kspace_jastrow.cpp
                test_short_range_cusp_jastrow.cpp ${MO_SRCS})
-TARGET_LINK_LIBRARIES(${UTEST_EXE} qmc qmcwfs qmcbase qmcutil ${QMC_UTIL_LIBS} ${MPI_LIBRARY})
+TARGET_LINK_LIBRARIES(${UTEST_EXE} qmcwfs qmcbase qmcutil ${QMC_UTIL_LIBS} ${MPI_LIBRARY})
 
 ADD_UNIT_TEST(${UTEST_NAME} "${QMCPACK_UNIT_TEST_DIR}/${UTEST_EXE}")
 SET_TESTS_PROPERTIES(${UTEST_NAME} PROPERTIES WORKING_DIRECTORY ${UTEST_DIR})
