+++ src/QMCWaveFunctions/tests/MinimalWaveFunctionPool.h
